backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Auth & security
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Migrations
alembic>=1.13.0

# RAG & uploads
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pypdf>=3.0.0
python-docx>=1.0.0
httpx>=0.25.0
python-multipart>=0.0.6